renamed folders and updated a test
