+++ training/unit-testing/go-lang/1.hello-world/README.md
@@ -1,0 +1,2 @@
+****See code examples in directory
+===========================
