+++ training/unit-testing/go-lang/1.hello-world/go.mod
@@ -1,0 +1,8 @@
+module github.com/ONSdigital/dp/training/unit-testing/go-lang/1.hello-world
+
+go 1.16
+
+require (
+	github.com/smartystreets/assertions v1.2.0
+	github.com/smartystreets/goconvey v1.6.4
+)
