+++ training/unit-testing/go-lang/3.dependency-injection/datastore/mock.go
@@ -4,12 +4,12 @@ import (
 	"github.com/stretchr/testify/mock"
 )
 
-//MockStore is a struct
+// MockStore is a struct
 type MockStore struct {
 	mock.Mock
 }
 
-//Get will get the int based on the ID
+// Get will get the int based on the ID
 func (m *MockStore) Get(ID int) (int, error) {
 	returnVals := m.Called(ID)
 	return returnVals.Get(0).(int), returnVals.Error(1)
